+++ src/grokcore/component/tests/adapter/nomodel.py
@@ -15,5 +15,6 @@ from zope import interface
 class IHome(interface.Interface):
     pass
 
+@grok.implementer(IHome)
 class Home(grok.Adapter):
-    grok.implements(IHome)
+    pass