+++ src/grokcore/component/tests/utility/implementsmany.py
@@ -17,5 +17,6 @@ class IClub(interface.Interface):
 class ISpikyClub(interface.Interface):
     pass
 
+@grok.implementer(IClub, ISpikyClub)
 class Club(grok.GlobalUtility):
-    grok.implements(IClub, ISpikyClub)
+    pass